handle fight now seperate
