+++ src/app.js
@@ -1,10 +1,11 @@
 "use strict";
 
+
 const domHandler = require("./domHandler");
 const buildRobot = require("./buildRobot");
-const attack = require("./attack");
+const handleFight = require("./handleFight");
 
-domHandler();
+domHandler.buildSelect();
 
 $("input").keypress( function(e) {
 	buildRobot.storeName(e, $(this));
@@ -15,5 +16,6 @@ $("select").change( function() {
 });
 
 $("#attack").click(() => {	
-	attack();
+	handleFight();
 });
+
